orchestrator/requirements.txt
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
supabase>=2.3.0
openai>=1.14.0
anthropic>=0.25.0
google-generativeai>=0.5.0
python-dotenv>=1.0.1
pydantic>=2.6.0
ollama>=0.1.8
